Modularize core, add comprehensive documentation, update changelog, and enforce project contribution guidelines
