+++ CHANGELOG.md
@@ -25,6 +25,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - GitHub Actions CI workflow: automatically builds, tests, lints (clippy), and checks formatting on push to main and on non-draft pull requests targeting main. Draft PRs are skipped.
 - Static compile-time projection check: Added a test to ensure that the projection from global to local session type for a role (e.g., Alice) is correct. This test is reported as a regular Rust test and fails to compile if the projection is incorrect.
 - Added docs/ImplementationOverview.md: a detailed overview and analysis of the current implementation, including goals, session type theory, combinator-by-combinator discussion, global/local types, compile-time properties, testing, and a comparison with other session type libraries (with direct links). Diagrams are properly separated for clarity.
+- Major codebase modularization: split core logic into `protocol.rs`, `types.rs`, and `introspection.rs` for clarity and maintainability.
+- Comprehensive doc comments for all major types, traits, and modules, following project documentation standards and cross-referencing advanced type-level patterns.
+- Top-level documentation for type-level map/fold patterns and the use of helper traits to resolve overlapping trait impls.
+- Improved documentation for protocol marker types, message primitives, and introspection traits.
+- Explicit user and Copilot contribution guidelines added to the project (see instructions section).
 
 ### Changed
 
@@ -47,6 +52,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Refactored combinator projection logic to use helper traits and avoid overlapping trait impls, improving maintainability and extensibility.
 - Updated documentation and README to include a dedicated section on projection from global to local session types.
 - All documentation files in `docs/` have been reformatted for markdownlint compliance: long lines wrapped, spacing and heading issues fixed, and style improved for readability and consistency. No content changes were made.
+- All protocol/session combinators, endpoint types, and projection traits are now re-exported at the crate root for user and test compatibility.
+- Code formatting and style updated to follow rustfmt and clippy recommendations.
 
 ### Removed
 
@@ -57,3 +64,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed all failing doctests by updating or removing outdated examples that used the old 4-argument form for `TInteract` and `TEnd`.
 - Resolved trait overlap and type equality issues in n-ary combinators and macros.
 - Ensured all tests compile and pass with the new pattern.
+- Resolved duplicate imports and module visibility issues after modularization.
+- All code and documentation pass `cargo check`, `cargo build`, `cargo fmt`, and `cargo clippy` (except for known/ignored test failures).
+
+### Known Issues
+
+- Some projection and trybuild tests (especially involving `TPar`/`EpPar`) are expected to fail due to ongoing design work. See protocol.rs and test files for details.
+
+---
